requirements.txt
google-genai
pydantic
python-dotenv
google-genai
pydantic
python-dotenv
